requirements.txt
# Document parsing
pypdfium2==5.5.0

# RAG pipeline
faiss-cpu==1.8.0
sentence-transformers==3.0.0
rank-bm25==0.2.2

# LLM API
openai==1.76.0

# App
streamlit==1.38.0

# Evaluation
pandas==2.2.2

# Config & env
pyyaml==6.0.2
python-dotenv==1.0.1

# Utils
numpy==1.26.4

# Logging
loguru==0.7.2

# Testing
pytest==8.3.5
